refactor: move InvalidDataException para um package mais generico por ser usado por mais de um service
