+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/InvalidDataException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions.despesa;
+package br.com.gestorfinanceiro.exceptions;
 
 public class InvalidDataException extends RuntimeException {
     public InvalidDataException(String message) {
